// src/app/model/wallet/register/user-create-req.ts
export class UserCreateRequest {
    type: string;
    firstName: string;
    lastName: string;
    email: string;
    password: string;
    retypePassword: string;
    address: string;
    city: string;
    state: string;
    postalCode: string;
    dateOfBirth: string;
    country: string;
    mobile: string;
    language: string;
    currency: string;
}